+++ src/plugin/built-in/props-propagation.js
@@ -7,7 +7,7 @@ module.exports = function(Doz, app) {
             // when update use this
             changes.forEach(change => {
                 if (
-                    change.type !== 'update' ||
+                    //change.type !== 'update' ||
                     (mainParent._propsPropagationIsArray && mainParent.propsPropagation.indexOf(change.currentPath) === -1)
                 ) return;
                 child.props[change.currentPath] = change.newValue;
@@ -71,7 +71,7 @@ module.exports = function(Doz, app) {
             component.propsPropagation = component.parent.propsPropagation;
             component._propsPropagationMainParent = component.parent._propsPropagationMainParent;
 
-            if (component.excludeFromPropsPropagation) {
+            if (component.excludeFromPropsPropagation || component.props['d:no-propagation'] !== undefined) {
                 Object.defineProperty(component, 'excludeFromPropsPropagation', {value: true})
             } else {
                 addToPropagation(component);