+++ test/doz.props-propagation.js
@@ -63,5 +63,58 @@ describe('Doz.props-propagation', function () {
 
 
         });
+
+        it('should be ok #2', function (done) {
+            document.body.innerHTML = `<div id="app"></div>`;
+
+            class SalutationCard extends Doz.Component{
+
+                constructor(o) {
+                    super(o);
+                    this.propsPropagation = true;
+                }
+
+                template(h) {
+                    return h`
+                        <${SalutationCardInner} d:no-propagation/>
+                    `
+                }
+
+                onMount() {
+                    this.props.name = 'Ipsum';
+                }
+            }
+
+            class SalutationCardInner extends Doz.Component{
+
+                constructor(o) {
+                    super(o);
+                }
+
+                template(h) {
+                    return h`
+                        <div>Hello ${this.props.name}</div>
+                    `
+                }
+
+            }
+
+            new Doz({
+                root: '#app',
+
+                template(h) {
+                    return h`
+                        <${SalutationCard}
+                            name="Doz">
+                        </>
+                    `
+                }
+            });
+
+            setTimeout(() => {
+                console.log(document.body.innerHTML);
+                be.err(done).equal(document.body.innerHTML, '<div id="app"><dz-app><salutation-card><salutation-card-inner><div>Hello undefined</div></salutation-card-inner></salutation-card></dz-app></div>')
+            }, 1000);
+        });
     });
 });